% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetter.R
\encoding{UTF-8}
\name{subsetter}
\alias{subsetter}
\title{Subset S4? object according to expression}
\usage{
subsetter(
  object.expr,
  inputAssay = "counts",
  minExpr = 1,
  minSamples = 2,
  across = "allSamples",
  groupBy = FALSE
)
}
\arguments{
\item{object.expr}{Expression object}

\item{inputAssay}{Options: c("counts", "tpm"). Default sets to "counts".}

\item{minExpr}{Minimum expression that would like to include in the subset condition. Default sets to 1.}

\item{minSamples}{Minimum library that would like to include in the subset condition. Default sets to 2.}

\item{across}{Condition that would like to subset including 
c("allSamples", "eachClass", "sumAllSamples", "sumInEachClass"). 
Default sets to "allSamples".}

\item{groupBy}{Column name of metadata. this will be set only for "eachClass", "sumInEachClass". 
Default sets to FALSE.}
}
\value{
new expression object according to the preferred condition
}
\description{
Subset S4? object according to expression
}
\examples{
FANTOM setting (Is it the FANTOM setting?, Robins setting) 
subset.expr <- subsetter(cage.expr, minExpr=2, minSamples=2, across="allSamples")
More than 1 tags in at least 2 libraies for at least 1 condition
subset.expr <- subsetter(cage.expr, minExpr=2, minSamples=2, across="eachClass", groupBy="Class")
More than 1 tags in at least 1 condition
subset.expr <- subsetter(cage.expr, minExpr=2, across="sumInEachClass", groupBy="Class")

}
